Proactively informing observers when shutting down (#15)
Rather than waiting for edge failure detection to kick in when a cluster has been shut down, this change proactively informs the observers of a node with a new Leaving message. In turn the observers the trigger edge failure alerting immediately.

Adds Cluster.leaveGracefully() and Cluster.shutdown() APIs for graceful and forced shutdowns respectfully. Accessing membership state after either of these APIs are invoked is illegal and will result in a thrown exception.


* Proactively informs observer nodes that the node is leaving when the cluster is shut down

* Leave notifications delivered in parallel, call to leave() protected by try/finally

* Fixing parallel leave message sending

- tolerating failure in delivering the messages, i.e. not cancelling other notifications
- adjusting test intervals in order to reach agreement faster

* Throw exceptions when trying to access membership state after shutting down