Endpoint performance and memory pressure optimization (#19)
This is a bit of a controversial change from the view point of the API, yet makes a lot of sense from the view point of performance and memory utilization for very large clusters.

The issue here is that the hostname of an Endpoint is modelled as a protobuf "string" type. This type carries with it the overhead of encoding to or decoding from UTF-8 every time a message is sent or received (and the field accessed).

From the view point of the algorithms in place, there's no added value in having the endpoint host data be encoded as byte array or utf-8 encoded string. It is just data, what matters is that the ordering of the endpoints can be established. Having a string only matters at the interfaces: when configuring a hostname, when sending a message to one and when printing log statements (most of which at DEBUG/TRACE level).

Yet at the moment, when adding a new endpoint to the membership ring(s), the following code runs:

```
  public java.lang.String getHostname() {
    java.lang.Object ref = hostname_;
    if (ref instanceof java.lang.String) {
      return (java.lang.String) ref;
    } else {
      com.google.protobuf.ByteString bs =
          (com.google.protobuf.ByteString) ref;
      java.lang.String s = bs.toStringUtf8();
      hostname_ = s;
      return s;
    }
  }

```

For freshly received messages containing Endpoints, this means running `toStringUtf8()`, which when there are many is quite expensive in terms of CPU and memory usage.

This PR does the following:

- use `bytes` rather than `string` to encode the hostname in protobuf
- adjust all interfaces - the Cluster APIs are actually (almost) not affected since they use the `HostAndPort` construct
- use the existing underlying / existing byte array when computing the hashcode of an Endpoint in `Utils.AddressComparator`
- getting rid of the mapping between `Map<String, Metadata>` and `Map<Endpoint, Metadata>` by representing the map as two lists in protobuf (keys and values)

On local tests with 1000 concurrent nodes joining, there's a 10% improvement in memory allocation and a 20% improvement in CPU usage of the stack starting at the `TreeSet.add` method (39% vs 58%).