Migrate subscriptions to a ClusterStatusChange API
Previously, all subscriptions were callbacks with the signature
BiConsumer<Long, List<NodeStatusChange>>. This API was used to
present a configuration ID and a corresponding delta to the
caller. Depending on the kind of ClusterEvent, the delta was
either the proposed delta (for VIEW_CHANGE_PROPOSAL events),
or the delta that was just applied (for VIEW_CHANGE events).

This commit changes the subscription callback signatures to
Consumer<ClusterStatusChange>. A ClusterStatusChange object
exposes the configuration ID and delta like before, but
also supplies the full list of Endpoints that are part
of the configuration ID.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>