Address ClusterTest flakiness due to ConcurrentModificationException
Utils.AddressComparator caches AddressComparators per seed value,
and further caches the computed hash functions for performance.

An oversight led to the Utils.AddressComparator.INSTANCES map that stores
 the per-seed AddressComparators being static, which in turn causes
the /inner/ AddressComparator instances to be unintentionally
shared across rapid nodes being simultaneously launched in ClusterTest.

As a result, during membership changes from tests in ClusterTest,
we end up with CMEs that silently get gobbled by the threads. This
has caused some of these tests to be flakey over the years without
a clear cause. In fact, this has been exacerbated with all the performance
improvements from #18, the use of the AddressComparators to sort proposals,
and with CI hardware getting better, and finally became reproducible!

Note that this bug *only* affects ClusterTest. The unintended sharing cannot happen
in practice given that no one runs multiple rapid instances on the same JVM.

The fix is to remove Utils.AddressComparator.INSTANCES, and
maintain AddressComparator caches per instance of MembershipView,
as it should be.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>